move adc class to own file
